Extract stopwatch functionality into custom hook
