+++ .prettierrc
@@ -1,0 +1,4 @@
+{
+  "tabWidth": 4,
+  "useTabs": false
+}
