+++ src/App.tsx
@@ -1,51 +1,24 @@
-import React, { useRef, useState } from 'react'
-import StopWatchButton from './StopWatchButton'
-import StopWatch from './StopWatch'
-import "./styles.css"
+import React from "react";
+import StopWatchButton from "./StopWatchButton";
+import StopWatch from "./StopWatch";
+import { useStopWatch } from "./hooks/use-stop-watch";
+import "./styles.css";
 
 const App = () => {
-    const intervalId = useRef(null); 
+    const { time, isRunning, start, stop, reset } = useStopWatch();
 
-    const [time, setTime] = useState(0);
-    const [isRunning, setIsRunning] = useState(false);
-
-    const handleStart = () => {
-        if (!isRunning) {
-            setIsRunning(true);
-            intervalId.current = setInterval(() => {
-                setTime((time) => time + 10)
-            }, 10);
-        }
-    }
-
-    const handleStop = () => {
-        if (isRunning) {
-            setIsRunning(false);
-            clearInterval(intervalId.current);
-        }
-    }
-
-    const handleReset = () => {
-        handleStop();
-        setTime(0);
-    }
-
-    return(
+    return (
         <div className="stop-watch-container">
             <StopWatch time={time} />
             <div className="buttons-container">
-                <StopWatchButton onClick={handleStart} isDisabled={isRunning}>
+                <StopWatchButton onClick={start} isDisabled={isRunning}>
                     Start
                 </StopWatchButton>
-                <StopWatchButton onClick={handleStop}>
-                    Stop
-                </StopWatchButton>
-                <StopWatchButton onClick={handleReset}>
-                    Reset
-                </StopWatchButton>
+                <StopWatchButton onClick={stop}>Stop</StopWatchButton>
+                <StopWatchButton onClick={reset}>Reset</StopWatchButton>
             </div>
         </div>
-    )
-}
+    );
+};
 
 export default App;
