+++ src/StopWatchButton.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react'
+import React, { ReactNode } from "react";
 
 interface StopWatchButtonProps {
     children?: ReactNode;
@@ -6,16 +6,20 @@ interface StopWatchButtonProps {
     onClick?: () => void;
 }
 
-const StopWatchButton = ({ children, isDisabled, onClick }: StopWatchButtonProps) => {
-    return(
-        <button 
-            className="stop-watch-button" 
-            disabled={isDisabled} 
+const StopWatchButton = ({
+    children,
+    isDisabled,
+    onClick,
+}: StopWatchButtonProps) => {
+    return (
+        <button
+            className="stop-watch-button"
+            disabled={isDisabled}
             onClick={onClick}
         >
             {children}
         </button>
-    )
-}
+    );
+};
 
 export default StopWatchButton;
